refactor: extract duplicate overlay code into mixin
